packages/scripts/requirements.txt
flask==2.2.2
Flask-RESTful==0.3.9
PyJWT==2.6.0

# Development
black==22.10.0
